+++ processor/src/test/resources/bad_input/com/example/observable/SetterButExpectSetterFalse2Model.java
@@ -1,0 +1,22 @@
+package com.example.observable;
+
+import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.Observable;
+
+@ArezComponent
+public class SetterButExpectSetterFalse2Model
+{
+  private long _field;
+
+  @Observable
+  public void setField( final long field )
+  {
+    _field = field;
+  }
+
+  @Observable( expectSetter = false )
+  public long getField()
+  {
+    return _field;
+  }
+}